Implement transaction deletion (#101)
* Rename FadeScrollView to ScrollableView

Rename `FadeScrollView` component, its associated CSS and TSX files to
`ScrollableView`. All usages of the component have been updated.

Also update `retend`, `retend-server`, and `retend-utils` packages to
their latest versions.

* Implement transaction deletion

- Refactor transaction sheet modes using a `Mode` query key - Introduce
`TransactionDeleteMode` component for confirmation - Add
`transactionDeleted` event and materializer for database - Enhance UX
with new `ToastMessage` and sheet animation constants

* Fix: Bump dependencies and various improvements

- Bump `@adbl/cells` and `@oxc-parser` dependencies - Fix: Use
contentRef to track the current sheet height for delete and edit actions
- Fix: Ensure `contentLoaded` state is reset when bottom sheet is closed
- Chore: Make sure there's no additional transition duration for
ExpandingView on exit - Fix: Don't remove the parent key of a subkey is
being removed. - Fix: Prevent the UI from becoming unmounted when
onboarding is finished - Fix: Prevent race condition from deleting a
transaction that has already been deleted - Fix: Update various
dependencies and runtime dependencies for `@oxc-parser`. - Fix: Ensure
Home Screen transitions correctly with the new `TransitionView` setup -
Fix: Use supportsCSSTypedOM to enable `attributeStyleMap` updates when
it's available.

* feat: Group transactions by date

Refactor the transaction listing to group transactions by date,
displaying each group with a date header. Utilize the `date-fns` library
for date formatting and manipulation. Update the relative time display
for transactions to show a time if it is not today.

* Refactor home page greeting

Extract the greeting component into its own file.

* feat: Add empty state to transaction listing

+++ utilities/composables/use-route-query-control.ts
@@ -1,4 +1,5 @@
 import { defer } from '@/utilities/miscellaneous'
+import { Cell } from 'retend'
 import { useRouteQuery, useRouter } from 'retend/router'
 
 export function useRouteQueryControl(_key: string | { _root: string }, value = 'true') {
@@ -10,18 +11,22 @@ export function useRouteQueryControl(_key: string | { _root: string }, value = '
    const add = () => {
       query.set(key, value)
    }
-   const remove = () => {
+   const remove = (options: { subKeys?: boolean } = { subKeys: true }) => {
       const subKeys = [...current.get().query.keys()].filter((_key) => {
          return _key.startsWith(`${key}.`)
       })
       query.delete(key)
-      defer(() => {
-         // This tries to prevent any subtle timing/rendering issues
-         // that can come from trying to remove all the keys at once.
-         query.delete(...subKeys)
-      })
+      if (options.subKeys) {
+         defer(() => {
+            // This tries to prevent any subtle timing/rendering issues
+            // that can come from trying to remove all the keys at once.
+            query.delete(...subKeys)
+         })
+      }
    }
-   const hasKey = query.has(key)
+   const hasKey = Cell.derived(() => {
+      return current.get().query.get(key) === value
+   })
 
    return { add, remove, hasKey }
 }
